Created a new directory apio/utils and moved to it the first file. More to follow.
