+++ apio/commands/apio_format.py
@@ -15,9 +15,10 @@ from typing import Tuple, List
 import click
 from click import secho
 from apio.apio_context import ApioContext, ApioContextScope
-from apio import pkg_util, util
+from apio import pkg_util
 from apio.commands import options
 from apio.managers import installer
+from apio.utils import util
 
 
 # ---------------------------
