+++ apio/commands/apio_graph.py
@@ -14,7 +14,7 @@ from apio.managers.scons import SCons
 from apio import cmd_util
 from apio.commands import options
 from apio.apio_context import ApioContext, ApioContextScope
-from apio.util import nameof
+from apio.utils.util import nameof
 
 
 # ---------------------------
