+++ apio/commands/apio_raw.py
@@ -15,7 +15,7 @@ from click import secho
 from apio import pkg_util, cmd_util
 from apio.apio_context import ApioContext, ApioContextScope
 from apio.commands import options
-from apio.util import nameof
+from apio.utils.util import nameof
 from apio.managers import installer
