+++ apio/utils/util.py
@@ -145,7 +145,7 @@ def get_path_in_apio_package(subpath: str) -> Path:
 
     # -- The parent folder is the apio root folder
     # -- Ex: /home/obijuan/.../site-packages/apio
-    path = current_python_file.parent
+    path = current_python_file.parent.parent
 
     # -- Add the given folder to the path. If subpath = "" this
     # -- does nothing, but fails if subpath is None.
